small refactor with AdminToggle
