+++ booking-UI/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@headlessui/react": "^1.7.18",
     "@tanstack/react-query": "^5.8.4",
     "axios": "^1.6.2",
     "class-variance-authority": "^0.7.0",
