+++ booking-UI/src/Pages/index.js
@@ -1,7 +1,7 @@
 import Book from "./Book";
 import Landing from "./Landing";
-import Login from "./Login";
-import Signup from "./Signup";
+import Login from "./Login/Login";
+import Signup from "./SignUp/Signup";
 import Users from "./Users";
 import Home from "./home";
